feeds/add: force cast category id to integer

+++ classes/feeds.php
@@ -1024,7 +1024,7 @@ class Feeds extends Handler_Protected {
 
 	function add(): void {
 		$feed = clean($_REQUEST['feed']);
-		$cat = clean($_REQUEST['cat'] ?? '');
+		$cat = (int) clean($_REQUEST['cat'] ?? '');
 		$need_auth = isset($_REQUEST['need_auth']);
 		$login = $need_auth ? clean($_REQUEST['login']) : '';
 		$pass = $need_auth ? clean($_REQUEST['pass']) : '';
